packages/cpbox/cpbox-1.5.16.tar.gz/cpbox-1.5.16/cpbox/tool/array.py
def chunks(l, n):
    """Yield successive n-sized chunks from l."""
    for i in range(0, len(l), n):
        yield l[i:i + n]

def merge(src, update):
    """
    >>> a = { 'first' : { 'all_rows' : { 'pass' : 'dog', 'number' : '1' } } }
    >>> b = { 'first' : { 'all_rows' : { 'fail' : 'cat', 'number' : '5' } } }
    >>> merge(a, b) == { 'first' : { 'all_rows' : { 'pass' : 'dog', 'fail' : 'cat', 'number' : '5' } } }
    True
    """
    for key, value in update.items():
        if isinstance(value, dict):
            # get node or create one
            node = src.setdefault(key, {})
            merge(node, value)
        else:
            src[key] = value

    return src
